another cpu generated pipeline added, tests reorganisation, a new CI for pipelines
