+++ .github/workflows/run_tests_pipelines_iris.yml
@@ -1,4 +1,4 @@
-name: HTTomo tests
+name: HTTomo pipelines tests
 
 on:
   pull_request:
@@ -32,15 +32,21 @@ jobs:
           post-cleanup: 'all'
           init-shell: bash
           
-      - name: Install httomo libraries and httomo
+      - name: Install httomo libraries, httomo-backends and httomo
         run: |
           micromamba activate httomo
           pip install --upgrade --force-reinstall pillow
           pip install httomolibgpu tomobar
           pip install --no-deps httomo-backends
-          pip install .         
+          pip install .   
           micromamba list
                 
-      - name: Run HTTomo tests
+      - name: Generate full yaml pipelines using pipeline directives
         run: |
-          pytest tests/
+          pip install "ruamel.yaml>0.18.0"
+          python ./docs/source/scripts/yaml_pipelines_generator.py -i ./docs/source/pipelines_full/gpu_pipeline1_directive.yaml -o ./docs/source/pipelines_full/gpu_pipeline1.yaml
+          python ./docs/source/scripts/yaml_pipelines_generator.py -i ./docs/source/pipelines_full/cpu_pipeline1_directive.yaml -o ./docs/source/pipelines_full/cpu_pipeline1.yaml
+
+      - name: Run HTTomo pipelines tests (small data)
+        run: |
+          pytest tests/test_pipeline_small.py --pipeline_small
