+++ docs/source/pipelines_full/cpu_pipeline1_directive.yaml
@@ -1,0 +1,18 @@
+- method: standard_tomo
+  module_path: httomo.data.hdf.loaders
+- method: find_center_vo
+  module_path: tomopy.recon.rotation
+- method: remove_outlier
+  module_path: tomopy.misc.corr
+- method: normalize
+  module_path: tomopy.prep.normalize
+- method: minus_log
+  module_path: tomopy.prep.normalize
+- method: recon
+  module_path: tomopy.recon.algorithm
+- method: calculate_stats
+  module_path: httomo.methods
+- method: rescale_to_int
+  module_path: httomolibgpu.misc.rescale
+- method: save_to_images
+  module_path: httomolib.misc.images
